Desktop/labaweb/website/requirements.txt
fastapi==0.112.3
sqlalchemy==2.0.35
pydantic==2.8.2
passlib==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
psycopg2-binary==2.9.10
jinja2==3.1.4
starlette==0.38.4  
PyJWT==2.3.0
websocket
itsdangerous==2.2.0
uvicorn[standard]
python-multipart==0.0.17